Signed-off-by: Nao YONASHIRO <yonashiro@r.recruit.co.jp>

improve performance.
```
name               old time/op  new time/op  delta
Create-4           13.1µs ± 2%   8.7µs ±10%  -33.99%  (p=0.000 n=9+10)
CreateBuildInfo-4   920ns ± 1%   422ns ±10%  -54.12%  (p=0.000 n=9+10)
```