@beorn7 2018-10-19
The problem with the `strings.Replacer` is that it causes allocs, see https://golang.org/src/strings/strings.go?s=22223:22269#L840

The runtime gain is significant, but allocations can bite us in different ways.

I think, however, most of the speed gain happens because most strings don't need escaping at all and could just be copied directly. Thus, we could try to use the same strategy as in `strings.Replacer`, i.e. first check if escaping is needed at all, and if not, just call `w.WriteString` directly. That should have an even better effect than the one in this PR.