@beorn7 2018-10-19
Sorry, misread `Replace` vs. `Replacer`.

I now think this is doing exactly the right thing. Will just test for allocs to be sure onec I'm at my development computer.