@beorn7 2018-10-19
Confirmed allocation-free-ness. This is exactly what I was looking for but failed to spot in the `scrings` package. Many thanks for this.

Compared to the original benchmark in #148, this is a reduction in runtime of 54% for `Create` and `67% for `CreateBuildInfo`. Really nice.